requirements.txt
requests
python-dateutil